+++ client/src/js/components/global/addThousandComma.js
@@ -5,6 +5,6 @@ function addThousandComma(number) {
   };
 
   return number.toFixed(2).replace(/\B(?=(\d{3})+(?!\d))/g, ",");
-}
+};
 
 export default addThousandComma;